requirements.txt
distro==1.5.0
et-xmlfile==1.1.0
numpy==1.20.2
openpyxl==3.0.7
pandas==1.2.4
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
tabula-py==2.2.0
Unidecode==1.2.0
xlrd==2.0.1
